Added traefik management for forward auth.

+++ device-microservice/demo/src/main/java/com/platform/device/controllers/DeviceController.java
@@ -2,8 +2,11 @@ package com.platform.device.controllers;
 
 import com.platform.device.dtos.DeviceDTO;
 import com.platform.device.dtos.UserDeviceDTO;
+import com.platform.device.entities.Role;
 import com.platform.device.services.DeviceService;
+import com.platform.device.services.SecurityService;
 import com.platform.device.services.UserDeviceService;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -22,23 +25,32 @@ public class DeviceController {
 
     @Autowired
     private UserDeviceService userDeviceService;
+    
+    @Autowired
+    private SecurityService securityService;
 
     @GetMapping("/get-all")
     @ResponseBody
-    public ResponseEntity<List<DeviceDTO>> getAllDevices(){
+    public ResponseEntity<List<DeviceDTO>> getAllDevices(HttpServletRequest request){
+        // USER and ADMIN can access
+        securityService.requireRole(request, Role.USER, Role.ADMIN);
         return ResponseEntity.ok(this.deviceService.getAllDevices());
     }
 
     @GetMapping
     @ResponseBody
-    public ResponseEntity<DeviceDTO> getDeviceById(@RequestParam UUID id){
+    public ResponseEntity<DeviceDTO> getDeviceById(@RequestParam UUID id, HttpServletRequest request){
+        // USER and ADMIN can access
+        securityService.requireRole(request, Role.USER, Role.ADMIN);
         DeviceDTO device = this.deviceService.getDeviceById(id);
         return ResponseEntity.ok(device);
     }
 
     @PostMapping
     @ResponseBody
-    public ResponseEntity<DeviceDTO> createDevice(@RequestBody DeviceDTO device){
+    public ResponseEntity<DeviceDTO> createDevice(@RequestBody DeviceDTO device, HttpServletRequest request){
+        // ADMIN only
+        securityService.requireAdmin(request);
         DeviceDTO createdDevice = this.deviceService.createDevice(device);
 
         URI location = ServletUriComponentsBuilder
@@ -51,14 +63,18 @@ public class DeviceController {
 
     @PutMapping
     @ResponseBody
-    public ResponseEntity<DeviceDTO> updateDevice(@RequestBody DeviceDTO device){
-        DeviceDTO newDevice = this.deviceService.updateUser(device);
+    public ResponseEntity<DeviceDTO> updateDevice(@RequestBody DeviceDTO device, HttpServletRequest request){
+        // ADMIN only
+        securityService.requireAdmin(request);
+        DeviceDTO newDevice = this.deviceService.updateDevice(device);
         return ResponseEntity.accepted().body(newDevice);
     }
 
     @DeleteMapping
     @ResponseBody
-    public ResponseEntity<String> deleteDeviceById(@RequestParam UUID id){
+    public ResponseEntity<String> deleteDeviceById(@RequestParam UUID id, HttpServletRequest request){
+        // ADMIN only
+        securityService.requireAdmin(request);
         this.deviceService.deleteById(id);
         return ResponseEntity.ok("Deletion Successful");
     }
@@ -66,13 +82,24 @@ public class DeviceController {
     @PostMapping("user-device")
     @ResponseBody
     // DTO only contains userId and deviceId
-    public UserDeviceDTO addDeviceForUser(@RequestBody UserDeviceDTO userDeviceDTO){
+    public UserDeviceDTO addDeviceForUser(@RequestBody UserDeviceDTO userDeviceDTO, HttpServletRequest request){
+        // USER and ADMIN can access
+        securityService.requireRole(request, Role.USER, Role.ADMIN);
+        if(securityService.getUserRole(request) == Role.USER) {
+            securityService.requireUser(request, userDeviceDTO.getUserId());
+        }
         return this.userDeviceService.createUserDevice(userDeviceDTO);
     }
 
     @DeleteMapping("user-device")
     @ResponseBody
-    public ResponseEntity<String> deleteDeviceForUser(@RequestParam UUID userId, @RequestParam UUID deviceId){
+    public ResponseEntity<String> deleteDeviceForUser(@RequestParam UUID userId, @RequestParam UUID deviceId, HttpServletRequest request) {
+        // USER and ADMIN can access
+        securityService.requireRole(request, Role.USER, Role.ADMIN);
+        if (securityService.getUserRole(request) == Role.USER){
+            securityService.requireUser(request, userId);
+
+        }
         this.userDeviceService.deleteUserDevice(userId, deviceId);
         return ResponseEntity.ok("Deletion Successful");
     }
@@ -80,7 +107,10 @@ public class DeviceController {
     // GET DEVICE FOR USER
     @GetMapping("user-device")
     @ResponseBody
-    public ResponseEntity<List<DeviceDTO>> getDevicesForUser(@RequestParam UUID userId){
+    public ResponseEntity<List<DeviceDTO>> getDevicesForUser(@RequestParam UUID userId, HttpServletRequest request){
+        // USER and ADMIN can access
+        securityService.requireRole(request, Role.USER, Role.ADMIN);
+        securityService.requireUser(request, userId);
         List<DeviceDTO> devices = this.userDeviceService.getDevicesForUser(userId);
         return ResponseEntity.ok(devices);
     }
